recognition/requirements.txt
PyYAML
nibabel
numpy
pillow
scikit-image
scikit-learn
scipy
torch
torchvision